requirements.txt
django==3.2.0
django-cors-headers
djangorestframework
wheel
djangorestframework-simplejwt
pillow
pyotp
psycopg2